+++ src/main/java/net/thumbtack/repo/mapper/ProductMapper.java
@@ -10,7 +10,7 @@ public class ProductMapper implements RowMapper<Product> {
   @Override
   public Product mapRow(ResultSet resultSet, int i) throws SQLException {
     Product product = new Product();
-    product.setId(resultSet.getInt("id"));
+    product.setId(resultSet.getLong("id"));
     product.setNameProduct(resultSet.getString("nameProduct"));
     product.setPrice(resultSet.getInt("price"));
     return product;
